united user related dtos in one block (user-dtos)
