+++ server/decorators/current-user.decorator.ts
@@ -1,5 +1,5 @@
 import { createParamDecorator, ExecutionContext } from "@nestjs/common";
-import { CurrentUserDTO } from "src/users/dtos/current-user.dto";
+import { CurrentUserDTO } from "src/users/dtos/user-dtos";
 
 export const User = createParamDecorator(
     (data: CurrentUserDTO, ctx: ExecutionContext)=>{
