+++ server/interfaces/user-service.interface.ts
@@ -1,12 +1,14 @@
 import mongoose from 'mongoose';
-import { UserDataDTO } from 'src/users/dtos/user-data.dto';
 import { ChatGroupInfoDTO } from 'src/chat-groups/dtos/chat-group-info.dto';
-import { CreateUserDTO } from 'src/users/dtos/create-user.dto';
-import { ReturnUserDTO } from 'src/users/dtos/return-user.dto';
 import { ReturnUserProfile } from 'src/users/users.model';
-import { FriendInfoDTO } from 'src/users/dtos/friend-info.dto';
-import { UserToBeValidateDTO } from 'src/users/dtos/user-tobe-validate.dto';
-import { UserProfileInfoDTO } from 'src/users/dtos/user-profile-info.dto';
+import { 
+    UserToBeValidateDTO,
+    FriendInfoDTO,
+    UserProfileInfoDTO,
+    UserDataDTO,
+    ReturnUserDTO,
+    CreateUserDTO
+     } from 'src/users/dtos/user-dtos';
 export interface IUsersService {
     createUser({createUserDTO} : {createUserDTO: CreateUserDTO }):Promise<ReturnUserDTO>;
     findUser({userId} : {userId:mongoose.Types.ObjectId}): Promise<UserProfileInfoDTO>;
@@ -16,7 +18,7 @@ export interface IUsersService {
     addFriend({userId, friendId} : {userId:mongoose.Types.ObjectId, friendId: mongoose.Types.ObjectId}) : Promise<FriendInfoDTO>;
     removeFriend({userId, friendId} : {userId:mongoose.Types.ObjectId, friendId: mongoose.Types.ObjectId}) : Promise<FriendInfoDTO>;
     getFriendIdsOfUser({userId} : {userId: mongoose.Types.ObjectId}) : Promise<mongoose.Types.ObjectId[]>;
-    getUserToBeValidate({id}:{id: mongoose.Types.ObjectId} ) : Promise<UserToBeValidateDTO>
+    getUserToBeValidate({userId}:{userId: mongoose.Types.ObjectId} ) : Promise<UserToBeValidateDTO>
     mapUserProfileInfo({userId, name, email, chatGroupDetails, friendsData} : {userId:mongoose.Types.ObjectId, name: string, email: string, chatGroupDetails: ChatGroupInfoDTO[], friendsData:UserDataDTO[]}) : Promise<ReturnUserProfile>;
     searchUser({searchText} : {searchText:string}): Promise<any>;
 }
