+++ server/src/auth/local.strategy.ts
@@ -2,7 +2,7 @@ import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { Strategy } from 'passport-local';
 import { AuthService } from './auth.service';
-import { LoginUserDTO } from 'src/users/dtos/login-user.dto';
+import { LoginUserDTO } from 'src/users/dtos/user-dtos';
 
 @Injectable()
 export class LocalStrategy extends PassportStrategy(Strategy) {
