+++ server/src/mapper/user-mapper.ts
@@ -2,10 +2,12 @@ import { Injectable, Logger } from "@nestjs/common";
 import { createMap, forMember, mapFrom, Mapper, MappingProfile } from '@automapper/core';
 import { AutomapperProfile, InjectMapper } from '@automapper/nestjs';
 import { ReturnUser, ReturnUserProfile, UserToBeValidate } from "src/users/users.model";
-import { ReturnUserDTO } from "src/users/dtos/return-user.dto";
-import { UserProfileInfoDTO } from "src/users/dtos/user-profile-info.dto";
-import { FriendInfoDTO } from "src/users/dtos/friend-info.dto";
-import { UserToBeValidateDTO } from "src/users/dtos/user-tobe-validate.dto";
+import { 
+    UserToBeValidateDTO,
+    UserProfileInfoDTO,
+    ReturnUserDTO,
+    FriendInfoDTO
+ } from "src/users/dtos/user-dtos";
 @Injectable()
 export class UserProfile extends AutomapperProfile{
     constructor(@InjectMapper() protected readonly mapper: Mapper){
