--- server/src/users/dtos/authenticated-user.dto.ts
@@ -1,19 +1,0 @@
-import { Expose, Exclude } from "class-transformer";
-import { IsEmpty } from "class-validator";
-export class AuthenticatedUserDTO {
-    @Exclude()
-    @IsEmpty({message:'Users must be authenticated, authentication is required !'})
-    access_token:string;
-    
-    @Expose()
-    @IsEmpty()
-    userId: string;
-    
-    @Expose()
-    @IsEmpty()
-    userEmail: string;
-    
-    @Expose()
-    @IsEmpty()
-    userName: string;
-}
