--- server/src/users/dtos/current-user.dto.ts
@@ -1,9 +1,0 @@
-import { Expose } from "class-transformer";
-export class CurrentUserDTO {
-    @Expose()
-    userId: number;
-    @Expose()
-    userEmail: string;
-    @Expose()
-    userName: string;
-}
