--- server/src/users/dtos/friend-info.dto.ts
@@ -1,7 +1,0 @@
-import mongoose from "mongoose";
-
-export class FriendInfoDTO{
-    _id: mongoose.Types.ObjectId;
-    name: string;
-    email: string;
-}
