--- server/src/users/dtos/user-data.dto.ts
@@ -1,8 +1,0 @@
-import mongoose from "mongoose";
-
-export class UserDataDTO {
-    _id: mongoose.Types.ObjectId;
-    name: string;
-    email: string;
-    ChatGroups: mongoose.Types.ObjectId[];
-}
