--- server/src/users/dtos/user-profile-info.dto.ts
@@ -1,9 +1,0 @@
-import mongoose from "mongoose";
-
-export class UserProfileInfoDTO {
-    UserId: mongoose.Types.ObjectId;
-    UserName: string;
-    UserEmail: string;
-    ChatGroups: any[];
-    Friends: any[];
-}
