+++ server/src/users/users.model.ts
@@ -2,7 +2,7 @@ import { Prop } from '@nestjs/mongoose';
 import * as mongoose from 'mongoose';
 import { Schema as mSchema, Document, Types } from 'mongoose';
 import { ChatGroupInfoDTO } from 'src/chat-groups/dtos/chat-group-info.dto';
-import { FriendInfoDTO } from './dtos/friend-info.dto';
+import { FriendInfoDTO } from './dtos/user-dtos';
 export interface User {
     _id:string; 
     name: string;
