+++ server/src/users/users.service.ts
@@ -1,16 +1,20 @@
 import { Injectable, Logger } from '@nestjs/common';
 import mongoose from 'mongoose';
 import { UsersRepository } from './users.repository';
-import { UserProfileInfoDTO } from './dtos/user-profile-info.dto';
 import { ChatGroupInfoDTO } from 'src/chat-groups/dtos/chat-group-info.dto';
 import { IUsersService } from 'interfaces/user-service.interface';
-import { CreateUserDTO } from './dtos/create-user.dto';
 import { InjectMapper } from '@automapper/nestjs';
 import { Mapper } from '@automapper/core';
-import { ReturnUser, ReturnUserProfile, UserToBeValidate } from './users.model';
-import { ReturnUserDTO } from './dtos/return-user.dto';
-import { FriendInfoDTO } from './dtos/friend-info.dto';
-import { UserToBeValidateDTO } from './dtos/user-tobe-validate.dto';
+import { 
+    ReturnUser, 
+    ReturnUserProfile, 
+    UserToBeValidate } from './users.model';
+import { 
+    UserToBeValidateDTO, 
+    UserProfileInfoDTO,
+    ReturnUserDTO,
+    FriendInfoDTO,
+    CreateUserDTO } from './dtos/user-dtos';
 
 @Injectable()
 export class UsersService implements IUsersService {
@@ -26,8 +30,11 @@ export class UsersService implements IUsersService {
             UserMapper,
             logger
         } = this;
+
         logger.debug(`[UsersService] createUser: ${JSON.stringify(createUserDTO)}`);
+
         const newUser : ReturnUser = await this.usersRepository.createUser({createUserDTO});
+
         return UserMapper.map< ReturnUser, ReturnUserDTO>(newUser,ReturnUser,ReturnUserDTO);
     }
     async findUser({userId}:{userId: mongoose.Types.ObjectId} ): Promise<UserProfileInfoDTO>{
@@ -36,21 +43,27 @@ export class UsersService implements IUsersService {
                 UserMapper,
                 logger
             } = this;
+
             logger.debug(`[UsersService] findUser: userId: ${JSON.stringify(userId)}`);
+
             const user = await this.usersRepository.findUserByObjectId(userId);
+
             return UserMapper.map<ReturnUser, UserProfileInfoDTO>(user, ReturnUser, UserProfileInfoDTO);
         } catch (error) {
             throw new Error(error);
         }
     }
-    async getUserToBeValidate({id}:{id: mongoose.Types.ObjectId} ): Promise<UserToBeValidateDTO>{
+    async getUserToBeValidate({userId}:{userId: mongoose.Types.ObjectId} ): Promise<UserToBeValidateDTO>{
         try {
             const {
                 UserMapper,
                 logger
             } = this;
-            logger.debug(`[UsersService] getUserToBeValidate: userId: ${JSON.stringify(id)}`)
-            const user = await this.usersRepository.findUserByObjectIdForValidating(id);
+
+            logger.debug(`[UsersService] getUserToBeValidate: userId: ${JSON.stringify(userId)}`)
+
+            const user = await this.usersRepository.findUserByObjectIdForValidating(userId);
+
             return UserMapper.map<UserToBeValidate, UserToBeValidateDTO>(user, UserToBeValidate, UserToBeValidateDTO);
         } catch (error) {
             throw new Error(error);
@@ -62,8 +75,11 @@ export class UsersService implements IUsersService {
                 UserMapper,
                 logger
             } = this;
+
             logger.debug(`[UsersService] findUserByEmail: email: ${JSON.stringify(email)}`);
+
             const user = await this.usersRepository.findByEmail(email);
+
             return UserMapper.map<ReturnUser, UserProfileInfoDTO>(user, ReturnUser, UserProfileInfoDTO);
         } catch (error) {
             throw new Error(error);
@@ -75,8 +91,11 @@ export class UsersService implements IUsersService {
                 UserMapper,
                 logger
             } = this;
+
             logger.debug(`[UsersService] addChatGroupToUser: userId: ${JSON.stringify(userId)}, chatGroupId: ${JSON.stringify(chatGroupId)}`);
+
             const user = await this.usersRepository.addChatGroupToUser(userId, chatGroupId);
+
             return UserMapper.map<ReturnUser, UserProfileInfoDTO>(user, ReturnUser, UserProfileInfoDTO);
         } catch (error) {
             throw new Error(error);
@@ -88,8 +107,11 @@ export class UsersService implements IUsersService {
                 UserMapper,
                 logger
             } = this;
+
             logger.debug(`[UsersService] removeChatGroupFromUser: userId: ${JSON.stringify(userId)}, chatGroupId: ${JSON.stringify(chatGroupId)}`);
+
             const user = await this.usersRepository.removeChatGroupFromUser(userId,chatGroupId);
+
             return UserMapper.map<ReturnUser, UserProfileInfoDTO>(user, ReturnUser, UserProfileInfoDTO);
         } catch (error) {
             throw new Error(error);
@@ -101,8 +123,11 @@ export class UsersService implements IUsersService {
                 UserMapper,
                 logger
             } = this;
+
             logger.debug(`[UsersService] addFriend: userId: ${JSON.stringify(userId)}, friendId: ${JSON.stringify(friendId)}`);
+
             const processedUser = await this.usersRepository.addFriend(userId, friendId);
+
             return UserMapper.map<ReturnUser,FriendInfoDTO>(processedUser, ReturnUser, FriendInfoDTO)
         } catch (error) {
             throw new Error(error);
@@ -114,8 +139,11 @@ export class UsersService implements IUsersService {
                 UserMapper,
                 logger
             } = this;
+
             logger.debug(`[UsersService] removeFriend: userId: ${JSON.stringify(userId)}, friendId: ${JSON.stringify(friendId)}`);
+
             const processedUser = await this.usersRepository.removeFriend(userId, friendId);
+            
             return UserMapper.map<ReturnUser,FriendInfoDTO>(processedUser, ReturnUser, FriendInfoDTO)
 
         } catch (error) {
@@ -125,8 +153,11 @@ export class UsersService implements IUsersService {
     async getFriendIdsOfUser( {userId} : {userId: mongoose.Types.ObjectId}): Promise<mongoose.Types.ObjectId[]> {
         try {
             const { logger } = this;
+
             logger.debug(`[UsersService] getFriendIdsOfUser: userId: ${JSON.stringify(userId)}`);
+
             return await this.usersRepository.getFriendsOfUser(userId);
+
         } catch (error) {
             throw new Error(error);
         }
@@ -137,11 +168,15 @@ export class UsersService implements IUsersService {
                 UserMapper,
                 logger
             } = this;
+
             logger.debug(`[UsersService] getUsersFriendsInfo: userIds: ${JSON.stringify(userIds)}`);
+
             const users = await this.usersRepository.getUserFriends(userIds);
+
             const usersData = Promise.all(users.map((user)=>{
                 return UserMapper.map<ReturnUser, FriendInfoDTO>(user, ReturnUser,FriendInfoDTO)
-            }))
+            }));
+
             return usersData;
         } catch (error) {
           throw new Error(error);
@@ -152,16 +187,20 @@ export class UsersService implements IUsersService {
             UserMapper,
             logger
         } = this;
+
         logger.debug(`[UsersService] mapUserProfileInfo:${JSON.stringify({userId, name, email, chatGroupDetails, friendsData})}`);
+
         const userProfileInfo = new UserProfileInfoDTO();
         userProfileInfo.UserId = userId;
         userProfileInfo.UserName = name;
         userProfileInfo.UserEmail = email;
         userProfileInfo.ChatGroups = chatGroupDetails;
         userProfileInfo.Friends = friendsData;
+
         return UserMapper.map<UserProfileInfoDTO, ReturnUserProfile>(userProfileInfo, UserProfileInfoDTO,ReturnUserProfile)
         
     }
+
     async searchUser({searchText}:{searchText:string}) {
         return await this.usersRepository.searchUser(searchText);
     }
